Fix Updated informataion when payment successfull

+++ index.js
@@ -99,6 +99,9 @@ async function run() {
       const email = req.params.email;
       const query = { email: email };
       const user = await usersCollection.findOne(query);
+      if (!user) {
+        return res.status(404).send({ error: true, message: 'User not found' });
+      }
       res.send(user);
     });
 
@@ -197,45 +200,75 @@ async function run() {
     // Selected Class
     app.post('/selectedClass', async (req, res) => {
       const selectedClass = req.body;
-      const query = { selectedClassId: selectedClass.selectedClassId }
+      const query = {
+        selectedClassId: selectedClass.selectedClassId,
+        email: selectedClass.email
+
+      };
+
       const existingSelectedClass = await selectedClassCollection.findOne(query);
       if (existingSelectedClass) {
-        return res.send({ message: 'Classes ALready Selected' })
+        return res.send({ message: 'You already selected this class' });
       }
+
       const result = await selectedClassCollection.insertOne(selectedClass);
       res.send(result);
-    })
+    });
     // Payment Related Api
     app.post('/payments', verifyJWT, async (req, res) => {
       try {
         const payment = req.body;
+        const existingPayment = await paymentClassCollection.findOne({ classId: payment.classId });
+    
+        if (existingPayment) {
+          return res.send({ message: 'You have already made a payment for this class' });
+        }
+    
         const insertResult = await paymentClassCollection.insertOne(payment);
-
+    
         const classId = payment.classId;
+        const selectedClassId = payment.selectedClassId;
+    
+        console.log(classId);
         const deleteResult = await selectedClassCollection.deleteOne({ _id: new ObjectId(classId) });
-
+    
         // Update enrollment status and available seats
         const updateClassResult = await addClassCollection.updateOne(
-          { _id: new ObjectId(classId) },
+          {
+            _id: new ObjectId(selectedClassId)
+          },
           {
             $inc: {
-              availableSeats: -1
-            }, $inc: { enrollmentStudent: 1 }
+              availableSeats: -1,
+              enrollmentStudent: 1
+            }
           }
         );
-
+    
         // Update totalEnrolmentStudent for instructor
-        const instructorEmail = payment.instructorEmail;
         const updateUserResult = await usersCollection.updateOne(
-          { email: instructorEmail },
+          { email: payment.instructorEmail },
           { $inc: { totalEnrolmentStudent: 1 } }
         );
-
+    
+        console.log('sldjflsdjfl',updateClassResult)
         res.send({ insertResult, deleteResult, updateClassResult, updateUserResult });
       } catch (error) {
         console.error('Error in processing payment:', error);
         res.status(500).send('Error in processing payment');
       }
+    });
+
+    //  GET Payment Api
+    app.get('/paymentHistory/:email', verifyJWT, async (req, res) => {
+      const { email } = req.params;
+      const query = { email: email };
+      const paymentHistory = await paymentClassCollection
+        .find(query)
+        .sort({ date: -1 })
+        .toArray();
+
+      res.send(paymentHistory);
     });
 
     //Instructor Approve Class
